improved wall filter in footer
