+++ api/giffygram.json
@@ -50,7 +50,7 @@
       "title": "My first post",
       "imageURL": "https://media.giphy.com/media/3oz8xyUoD2HlTIcdTW/giphy.gif",
       "description": "Jianbing butcher post-ironic brunch marfa, listicle quinoa kickstarter vinyl poke cornhole snackwave yr flexitarian.",
-      "timestamp": 2021
+      "timestamp": 1508080619498
     },
     {
       "id": 2,
@@ -58,7 +58,7 @@
       "title": "Dinner",
       "imageURL": "https://media.giphy.com/media/4RslNp4SUqAoM/giphy.gif",
       "description": "Messenger bag hella gochujang adaptogen, aesthetic sartorial before they sold out franzen sustainable thundercats.",
-      "timestamp": 2020
+      "timestamp": 1497580619498
     },
     {
       "id": 3,
@@ -66,14 +66,14 @@
       "title": "Graduation!",
       "imageURL": "https://media.giphy.com/media/lSPlEENLTonvclZP44/giphy.gif",
       "description": "Pop-up biodiesel pickled crucifix. Tousled succulents banjo bicycle rights.",
-      "timestamp": 2019
+      "timestamp": 1528720619498
     },
     {
       "title": "Learning JavaScript",
       "imageURL": "https://media.giphy.com/media/S9dN7OWFj8GoRhTIuL/giphy-downsized.gif",
       "description": "Ethical chillwave jianbing ramps plaid subway tile.",
       "userId": 2,
-      "timestamp": 2018,
+      "timestamp": 1617396786968,
       "id": 4
     },
     {
@@ -81,7 +81,7 @@
       "imageURL": "",
       "description": "",
       "userId": 1,
-      "timestamp": 2020,
+      "timestamp": 1621534541361,
       "id": 5
     }
   ],
@@ -145,6 +145,11 @@
       "postId": 4,
       "userId": 4,
       "id": 12
+    },
+    {
+      "postId": 3,
+      "userId": 4,
+      "id": 13
     }
   ],
   "messages": [
